Re-organize libsubmit subdirectories
Note that libsubmit tests were not being run by travis; I'll open an issue to
integrate them separately. For now I've just dumped them in `tests/integration`,
which is ignored by pytest.
