--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "parsl/libsubmit"]
-	path = parsl/libsubmit
-	url = https://github.com/Parsl/libsubmit
