--- libsubmit/tests/setup_path.sh
@@ -1,2 +1,0 @@
-export PATH=$PWD:$PATH
-export PYTHONPATH=$PWD/../../:$PYTHONPATH
