--- libsubmit/version.py
@@ -1,4 +1,0 @@
-''' Set module version
-<Major>.<Minor>.<maintenance>[-alpha/beta/..]
-'''
-VERSION = '0.5.1-a3'
