+++ parsl/providers/aws/template.py
