+++ parsl/providers/cobalt/template.py
