+++ parsl/providers/condor/template.py
