+++ parsl/providers/grid_engine/template.py
