+++ parsl/providers/jetstream/setup_first_time.sh
